Upgrade jetty to recent version. Update to java 1.8 (required for jetty)
